+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/whatsnew/WhatsNewViewModel.kt
@@ -7,11 +7,7 @@ import au.com.shiftyjelly.pocketcasts.models.to.SubscriptionStatus
 import au.com.shiftyjelly.pocketcasts.models.type.Subscription.SubscriptionTier
 import au.com.shiftyjelly.pocketcasts.models.type.SubscriptionFrequency
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
-import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
-import au.com.shiftyjelly.pocketcasts.repositories.subscription.SubscriptionManager
 import au.com.shiftyjelly.pocketcasts.settings.onboarding.OnboardingUpgradeSource
-import au.com.shiftyjelly.pocketcasts.utils.earlyaccess.EarlyAccessStrings
-import au.com.shiftyjelly.pocketcasts.utils.featureflag.BookmarkFeatureControl
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureFlag
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.UserTier
@@ -27,10 +23,7 @@ import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @HiltViewModel
 class WhatsNewViewModel @Inject constructor(
-    private val playbackManager: PlaybackManager,
-    private val subscriptionManager: SubscriptionManager,
     private val settings: Settings,
-    private val bookmarkFeature: BookmarkFeatureControl,
 ) : ViewModel() {
     private val _state: MutableStateFlow<UiState> = MutableStateFlow(UiState.Loading)
     val state = _state.asStateFlow()
@@ -41,7 +34,9 @@ class WhatsNewViewModel @Inject constructor(
     init {
         viewModelScope.launch {
             settings.cachedSubscriptionStatus.flow.stateIn(viewModelScope).collect { subscriptionStatus ->
-                if (FeatureFlag.isEnabled(Feature.SLUMBER_STUDIOS_YEARLY_PROMO)) {
+                if (FeatureFlag.isEnabled(Feature.DESELECT_CHAPTERS)) {
+                    updateStateForDeselectChapters()
+                } else if (FeatureFlag.isEnabled(Feature.SLUMBER_STUDIOS_YEARLY_PROMO)) {
                     val previousUserTier = (_state.value as? UiState.Loaded)?.tier
                     // Close what's new if plan changed but user is not eligible for slumber studios promo
                     // This is done so that user doesn't subscribe from the what's new to yearly plan now and gets subscribed to two plans
@@ -52,8 +47,6 @@ class WhatsNewViewModel @Inject constructor(
                     } else {
                         updateStateForSlumberStudiosPromo(subscriptionStatus)
                     }
-                } else {
-                    updateStateForBookmarks()
                 }
             }
         }
@@ -80,51 +73,6 @@ class WhatsNewViewModel @Inject constructor(
                 tier = userTier,
             )
         }
-    }
-
-    private fun updateStateForBookmarks() {
-        val userTier = settings.userTier
-        val isUserEntitled = bookmarkFeature.isAvailable(userTier)
-        if (isUserEntitled) {
-            _state.value = UiState.Loaded(
-                feature = bookmarksFeature(isUserEntitled = true),
-                tier = userTier,
-            )
-        } else {
-            viewModelScope.launch {
-                val subscriptionTier = SubscriptionTier.PLUS
-
-                subscriptionManager
-                    .freeTrialForSubscriptionTierFlow(subscriptionTier)
-                    .stateIn(viewModelScope)
-                    .collect { freeTrial ->
-                        _state.value = UiState.Loaded(
-                            feature = bookmarksFeature(
-                                trialExists = freeTrial.exists,
-                                isUserEntitled = false,
-                                subscriptionTier = subscriptionTier,
-                            ),
-                            tier = userTier,
-                        )
-                    }
-            }
-        }
-    }
-
-    private fun bookmarksFeature(
-        trialExists: Boolean = false,
-        isUserEntitled: Boolean,
-        subscriptionTier: SubscriptionTier? = null,
-    ): WhatsNewFeature.Bookmarks {
-        val currentEpisode = playbackManager.getCurrentEpisode()
-        return WhatsNewFeature.Bookmarks(
-            title = EarlyAccessStrings.getAppropriateTextResource(LR.string.whats_new_bookmarks_title),
-            message = if (isUserEntitled) LR.string.whats_new_bookmarks_body else LR.string.bookmarks_upsell_instructions,
-            confirmButtonTitle = if (currentEpisode == null) LR.string.whats_new_bookmarks_enable_now_button else LR.string.whats_new_bookmarks_try_now_button,
-            hasOffer = trialExists,
-            isUserEntitled = isUserEntitled,
-            subscriptionTier = subscriptionTier,
-        )
     }
 
     private fun updateStateForDeselectChapters() {
@@ -178,18 +126,7 @@ class WhatsNewViewModel @Inject constructor(
     fun onConfirm() {
         viewModelScope.launch {
             val currentState = state.value as? UiState.Loaded ?: return@launch
-            val currentEpisode = playbackManager.getCurrentEpisode()
             val target = when (currentState.feature) {
-                is WhatsNewFeature.Bookmarks -> if (currentState.feature.isUserEntitled) {
-                    if (currentEpisode == null) {
-                        NavigationState.HeadphoneControlsSettings
-                    } else {
-                        NavigationState.FullScreenPlayerScreen
-                    }
-                } else {
-                    NavigationState.StartUpsellFlow(OnboardingUpgradeSource.BOOKMARKS)
-                }
-
                 is WhatsNewFeature.SlumberStudiosPromo -> if (currentState.feature.isUserEntitled) {
                     NavigationState.SlumberStudiosRedeemPromoCode
                 } else {
@@ -232,19 +169,6 @@ class WhatsNewViewModel @Inject constructor(
         abstract val hasOffer: Boolean
         abstract val isUserEntitled: Boolean
         abstract val subscriptionTier: SubscriptionTier? // To show subscription when user is not entitled to the feature
-
-        data class Bookmarks(
-            @StringRes override val title: Int,
-            @StringRes override val message: Int,
-            @StringRes override val confirmButtonTitle: Int,
-            override val hasOffer: Boolean,
-            override val isUserEntitled: Boolean,
-            override val subscriptionTier: SubscriptionTier? = null,
-        ) : WhatsNewFeature(
-            title = title,
-            message = message,
-            confirmButtonTitle = confirmButtonTitle,
-        )
 
         data class SlumberStudiosPromo(
             val promoCode: String = "",